+++ src/main/java/com/demo/pteam/authentication/repository/entity/SocialAccountEntity.java
@@ -2,9 +2,9 @@ package com.demo.pteam.authentication.repository.entity;
 
 import com.demo.pteam.authentication.repository.converter.AccountStatusConverter;
 import com.demo.pteam.authentication.repository.converter.SocialTypeConverter;
-import com.demo.pteam.authentication.service.domain.AccountStatus;
-import com.demo.pteam.authentication.service.domain.Role;
-import com.demo.pteam.authentication.service.domain.SocialType;
+import com.demo.pteam.authentication.domain.AccountStatus;
+import com.demo.pteam.authentication.domain.Role;
+import com.demo.pteam.authentication.domain.SocialType;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
