+++ src/main/java/com/demo/pteam/chat/domain/Chat.java
@@ -1,0 +1,4 @@
+package com.demo.pteam.chat.domain;
+
+public class Chat {
+}
