--- src/main/java/com/demo/pteam/chat/service/domain/Chat.java
@@ -1,4 +1,0 @@
-package com.demo.pteam.chat.service.domain;
-
-public class Chat {
-}
