+++ src/main/java/com/demo/pteam/community/domain/Community.java
@@ -1,0 +1,4 @@
+package com.demo.pteam.community.domain;
+
+public class Community {
+}
