--- src/main/java/com/demo/pteam/community/service/domain/Community.java
@@ -1,4 +1,0 @@
-package com.demo.pteam.community.service.domain;
-
-public class Community {
-}
