+++ src/main/java/com/demo/pteam/credit/domain/CreditInfo.java
@@ -1,4 +1,4 @@
-package com.demo.pteam.credit.service.domain;
+package com.demo.pteam.credit.domain;
 
 public enum CreditInfo {
     // TODO: DB ENUM 바꾸기
