+++ src/main/java/com/demo/pteam/member/domain/Member.java
@@ -1,0 +1,4 @@
+package com.demo.pteam.member.domain;
+
+public class Member {
+}
