--- src/main/java/com/demo/pteam/member/service/domain/Member.java
@@ -1,4 +1,0 @@
-package com.demo.pteam.member.service.domain;
-
-public class Member {
-}
