+++ src/main/java/com/demo/pteam/notification/domain/Notification.java
@@ -1,0 +1,4 @@
+package com.demo.pteam.notification.domain;
+
+public class Notification {
+}
