--- src/main/java/com/demo/pteam/notification/service/domain/Notification.java
@@ -1,4 +1,0 @@
-package com.demo.pteam.notification.service.domain;
-
-public class Notification {
-}
