+++ src/main/java/com/demo/pteam/payment/domain/Payment.java
@@ -1,0 +1,4 @@
+package com.demo.pteam.payment.domain;
+
+public class Payment {
+}
