--- src/main/java/com/demo/pteam/payment/service/domain/Payment.java
@@ -1,4 +1,0 @@
-package com.demo.pteam.payment.service.domain;
-
-public class Payment {
-}
