+++ src/main/java/com/demo/pteam/review/domain/ReviewDomain.java
@@ -1,0 +1,4 @@
+package com.demo.pteam.review.domain;
+
+public class ReviewDomain {
+}
