--- src/main/java/com/demo/pteam/review/service/domain/ReviewDomain.java
@@ -1,4 +1,0 @@
-package com.demo.pteam.review.service.domain;
-
-public class ReviewDomain {
-}
