+++ src/main/java/com/demo/pteam/schedule/domain/Schedule.java
@@ -1,0 +1,4 @@
+package com.demo.pteam.schedule.domain;
+
+public class Schedule {
+}
