--- src/main/java/com/demo/pteam/schedule/service/domain/Schedule.java
@@ -1,4 +1,0 @@
-package com.demo.pteam.schedule.service.domain;
-
-public class Schedule {
-}
