+++ src/main/java/com/demo/pteam/trainer/domain/Trainer.java
@@ -1,0 +1,4 @@
+package com.demo.pteam.trainer.domain;
+
+public class Trainer {
+}
