--- src/main/java/com/demo/pteam/trainer/service/domain/Trainer.java
@@ -1,4 +1,0 @@
-package com.demo.pteam.trainer.service.domain;
-
-public class Trainer {
-}
