+++ src/main/java/com/demo/pteam/workout/domain/Workout.java
@@ -1,0 +1,4 @@
+package com.demo.pteam.workout.domain;
+
+public class Workout {
+}
