--- src/main/java/com/demo/pteam/workout/service/domain/Workout.java
@@ -1,4 +1,0 @@
-package com.demo.pteam.workout.service.domain;
-
-public class Workout {
-}
